+++ scripts/compelete_liquidity_provision.js
@@ -1,4 +1,4 @@
-const { deployFleetOfSafes } = require("./trading_strategy_helpers")
+const { deployFleetOfSafes } = require("./utils/trading_strategy_helpers")
 const Contract = require("@truffle/contract")
 const {
   getBundledTransaction,
