+++ scripts/utils/internals.js
