+++ scripts/wrap_eth.js
@@ -1,6 +1,6 @@
 const Contract = require("@truffle/contract")
 
-const { signAndSend, promptUser } = require("./sign_and_send")
+const { signAndSend, promptUser } = require("./utils/trading_strategy_helpers")
 const { CALL } = require("./trading_strategy_helpers")
 
 const argv = require("yargs")
